refactor(ir): retire load.ts; thin metadata-only ProgramType (D3-b.2 step 2)
Splits the retired `compiler/ir/load.ts` into two cleaner modules:

- `compiler/ir/lower_to_exprnode.ts` (~200 LOC) — `resolvedToSlotted`,
  the legacy ExprNode-shape lowerer used by `compile_resolved.ts` to
  feed `emit_numeric`. The eventual §2.1 from-scratch port of
  emit_numeric to walk ResolvedExpr directly would let us delete this
  too; for now it's the bridge.
- `compiler/ir/program_type_builder.ts` (~110 LOC) — `resolvedToProgramType`,
  builds a *metadata-only* ProgramType from a post-strata
  ResolvedProgram. typeName, port names, port types, register port
  types, rawInputDefaults — that's it. The heavy ExprNode-bearing
  fields on `ProgramDef` (outputExprNodes, registerExprNodes,
  delayUpdateNodes, nestedCalls, registerInitValues, delayInitValues,
  inputDefaults) are populated as empty arrays. Verified nothing
  external reads them.

Deletions:
- compiler/ir/load.ts (475 LOC) — `loadProgramDefFromResolved` was
  the bridge that built a slot-indexed ProgramDef carrying ExprNode
  trees for `flatten.ts` to assemble. Both consumers gone now.
- compiler/ir/load.test.ts (236 LOC) — unit tests for the deleted
  loader.

`compileResolvedToProgramDef` (in strata.ts) now drops its `session`
parameter (the retired loader needed it; the new builder doesn't).
Updated 4 callers in program.ts, session.ts, stdlib_loader.ts.

Tests: 814 pass, 1 skip, 1 local-only fail (the same untracked
schema_audit noise). Net delta: -742 LOC code, +310 LOC.

Remaining D3 work: trim ProgramDef interface to drop the now-empty
heavy fields (program_types.ts cleanup); reduce ProgramType /
ProgramInstance to thinner wrappers if useful.
